added a lot of tests and API code

+++ Touchcode/FakeTouchPoint.cs
@@ -5,7 +5,7 @@ namespace WpfApplication4.Touchcode
 {
     class FakeTouchPoint : TouchPoint
     {
-        public FakeTouchPoint(int x, int y)
+        public FakeTouchPoint(double x, double y)
             : base(new IHaveNoTouchDevice(666), new System.Windows.Point(x, y), new Rect(), new TouchAction())
         {
